Renamed openst_deployer to openst.
